Changing luma doubling test to a real anime frame.
Aoko.png was a bit unrealistic and I feel like this captures the subtle difference
between these shaders better on real video content. If you're wondering whether this is
unfair because ArtCNN is trained on Hibike, please remember only Chikai and Ensemble are used.
This frame is from the third anime season, and it was extracted from the shitty web source.

The compression artifacts also make this comparison more realistic as this is what most people
are ultimately watching.